configure db for heroku

+++ db.js
@@ -1,26 +1,23 @@
 const mysql = require('mysql2');
 
-// Create a MySQL connection with hard-coded credentials for local testing
-const connection = mysql.createConnection({
-  host     : process.env.CLEARDB_DATABASE_URL || 'localhost',
-  user     : 'songlibrarytest',
-  password : 'testtest',
-  multipleStatements: true,
-  dateStrings: true,
-  timezone: '+00:00'
-});
+let connection;
+let createCommand;
+let seedCommand;
 
-/*
-  Connect to the database described in the connection variable
-  Run commands to drop/create the database and song table and seed song data
-*/
-exports.connect = () => {
+// prod
+if(process.env.CLEARDB_DATABASE_URL) {
+  connection = mysql.createConnection({
+    host     : process.env.CLEARDB_DATABASE_URL,
+    user     : 'songlibrarytest',
+    password : 'testtest',
+    database : 'heroku_10d242b278d75f1',
+    multipleStatements: true,
+    dateStrings: true,
+    timezone: '+00:00'
+  });
 
   // command to create the song database and song table schema once
-  let createCommand = `
-    CREATE DATABASE IF NOT EXISTS burlinson_song_library;
-    use burlinson_song_library;
-
+  createCommand = `
     DROP TABLE song;
     CREATE TABLE song (
       id int(11) NOT NULL AUTO_INCREMENT PRIMARY KEY,
@@ -30,17 +27,50 @@ exports.connect = () => {
       price decimal(9,2) NOT NULL
     );
     `
-  // command to insert multiple songs at once
-  let seedCommand = `
-    INSERT INTO song (title, artist, release_date, price)
-    VALUES
-      ('Song1', 'Artist1', Date("2021-05-04"), 12.99),
-      ('Song2', 'Artist2', Date("2021-04-03"), 35.00),
-      ('Song3', 'Artist3', Date("2021-02-01"), 2.45),
-      ('Song4', 'Artist4', Date("2020-12-11"), 108.88),
-      ('Song5', 'Artist5', Date("2020-10-09"), 15.01)
-    ;
-    `
+
+  // local
+  } else {
+    connection = mysql.createConnection({
+      host     : 'localhost',
+      user     : 'songlibrarytest',
+      password : 'test',
+      multipleStatements: true,
+      dateStrings: true,
+      timezone: '+00:00'
+    });
+
+    createCommand = `
+      CREATE DATABASE IF NOT EXISTS burlinson_song_library;
+      use burlinson_song_library;
+
+      DROP TABLE song;
+      CREATE TABLE song (
+        id int(11) NOT NULL AUTO_INCREMENT PRIMARY KEY,
+        title varchar(100) NOT NULL,
+        artist varchar(100) NOT NULL,
+        release_date DATE NOT NULL,
+        price decimal(9,2) NOT NULL
+      );
+      `
+  }
+
+  /*
+    Connect to the database described in the connection variable
+    Run commands to drop/create the database and song table and seed song data
+  */
+  exports.connect = () => {
+
+    // command to insert multiple songs at once
+    seedCommand = `
+      INSERT INTO song (title, artist, release_date, price)
+      VALUES
+        ('Song1', 'Artist1', Date("2021-05-04"), 12.99),
+        ('Song2', 'Artist2', Date("2021-04-03"), 35.00),
+        ('Song3', 'Artist3', Date("2021-02-01"), 2.45),
+        ('Song4', 'Artist4', Date("2020-12-11"), 108.88),
+        ('Song5', 'Artist5', Date("2020-10-09"), 15.01)
+      ;
+      `
 
   // connect to the database and run the create and seed commands
   // throw an error and print the stack trace if the queries fail
